Move message processor into main app
Given our decision to proceed with keeping the API and sync worker as a
monolithic app for simplicity and to better fit in with the rest of the
GOV.UK ecosystem, this is the first step towards migrating away from the
semi-separated `DocumentSyncWorker` library.

- Move `DocumentSyncWorker::MessageProcessor` to the main app as
  `PublishingApiMessageProcessor`
- Run message processor directly from Rake task and get rid of
  additional `DocumentSyncWorker.run` abstraction
- Remove separate configuration class for `DocumentSyncWorker` and
  specify dependencies at point of use (and use Rails logger throughout)
