--- lib/document_sync_worker/configuration.rb
@@ -1,9 +1,0 @@
-module DocumentSyncWorker
-  class Configuration
-    attr_accessor :logger, :message_queue_name, :repository
-
-    def initialize
-      @logger = Rails.logger
-    end
-  end
-end
